tienda_pruebas/pyproject.toml
[project]
name = "tienda-pruebas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["django-payments>=2.0,<4.0"]
